Use calendar minimums
My first idea to filter out unrealistic calendar possibilities was to
use a "max_diff" meaning the farthest apart that the count of full and
hollow months could be. A max_diff of 4 worked pretty well for the
festival calendar but couldn't be applied to prytanies at all.

This adds calculation for the minimum count of full and hollow (or
long and short). These minimums are passed to the functions
calculating month combinations and ensure that there are enough months
left in the following months to bring the count of full and hollow in
the preceding up to a minimum.

max_diff will probably be removed in the future.