Delay loading data
Many functions have an optional data parameter which by default was
the return value of calling the load_data() function. This function
call would be resolved at the earliest opportunity and the resulting
dict passed dow to other functions. While there did not seem to be any
performance issue with passing around a large dict in a lot of nested
function calls, this did make the online help and pydoc output
unreadable, since the parameter was replaced with the full data in all
the documentation.

Replace the data parameter default values with the function rather
than the value of calling the function. Delay calling the function as
late as possible and add __optionally_load_data() as a way to get the
value of calling the function if a function has been passed, but just
use the value if it is not callable.

In private functions, make the parameter non-optional without a
default. Every public method that calls one of these private methods
must now resolve the function call if necessary.