Change quasi-solar cutoff to 376
Change the quasi-solar conciliar year cutoff date to 376 BCE. We can
say the quasi-solar year was in-use in 407/406 (IG I³ 377) but not in
346/345 (IG XII,6 1.261), but we have no direct evidence of when the
change was made. That this occurred in the restoration of democracy in
403, after the Thirty Tyrants, is a fine theory, but many arguments
have been made about the the calendar around the time of The 400 that
have proven false.

376 is the midpoint between the two dates and is a perfectly good
Bayesian prior. Aristotle does not seem to have know the quasi-solar
conciliar year, and he arrived in Athens around 366.