Add Macedonian option, Fewer default option values
Change how calendar options are handled. Some default values are
removed and values are only passed to the underlying festival calendar
functions of the options are set. This is is the default values of the
functions aren't overriden by default values set for the command-line
arguments.

Add Macedon as a calendar option